+++ Whiskey/Tasks/PublishBitbucketServerTag.ps1
@@ -21,6 +21,13 @@ function Publish-WhiskeyBBServerTag
             CredentialID: BitbucketServerCredential
             Uri: https://bitbucketserver.example.com'
 
+    if( $TaskContext.BuildMetadata.IsPullRequest )
+    {
+        'Skipping PublishBitbucketServerTag task: can''t tag a pull request commit because it doesn''t exist in the ' +
+        'origin repostory, only on the build server.' | Write-WhiskeyVerbose
+        return
+    }
+
     if( -not $TaskParameter['CredentialID'] )
     {
         Stop-WhiskeyTask -TaskContext $TaskContext -Message "Property 'CredentialID' is mandatory. It should be the ID of the credential to use when connecting to Bitbucket Server: